Add potential feature checker that will be for Resonant Inelastic X-ray Scattering (RIXS)
RIXS uses a diffraction grating to disperse a reflected X-ray beam onto a detector. To
super-resolve the locations of photons incident on the detector, a number of detector
parameters are required to discriminate single and multiple photon events at per-pixel
level. This package checks for the presence of those parameters.